my_spare function working

+++ csr.c
@@ -7,16 +7,15 @@ struct Flag {
 
 
 csr intoCSR(unsigned int size, double *mat){
-	unsigned int i, j=0, z=0, msize = size*size;
+	int i, j=0, z=0, msize = size*size;
 	csr m;
 	struct Flag f;
 
 	f.flag=1;
 
-	//m.ros = (size_t) size+1;
 	m.values = (double *) malloc(size * size * sizeof(double));
-	m.row_offsets = (size_t *) malloc( (size+1) * sizeof(size_t));
-	m.column_indices = (size_t *) malloc( size * size * sizeof(size_t));
+	m.row_offsets = (int *) malloc( (size+1) * sizeof(int));
+	m.column_indices = (int *) malloc( size * size * sizeof(int));
 
 	for(i=0; i < msize; i++){
 		if (i%size == 0){
@@ -28,7 +27,7 @@ csr intoCSR(unsigned int size, double *mat){
 			}
 			f.flag = 1;
 		}
-		if(mat[i] != 0){
+		if(mat[i] != 0.0){
 			m.values[j] = mat[i];
 			m.column_indices[j] = i%size;
 			if (f.flag == 1){
@@ -44,7 +43,8 @@ csr intoCSR(unsigned int size, double *mat){
 
 	//Resize vectors
 	m.values = (double *) realloc(m.values, j * sizeof(double));
-	m.column_indices = (size_t *) realloc(m.column_indices, j * sizeof(size_t));
+	m.column_indices = (int *) realloc(m.column_indices, j * sizeof(int));
+	//The last elemente is the size of the values vector
 	m.row_offsets[z] = j;
 
 	return m;
