+++ csr.h
@@ -1,11 +1,10 @@
 #ifndef CSR_MATRIX_H
 #define CSR_MATRIX_H
-#include <stddef.h>
 
 struct csr{
-    size_t *row_offsets;
+    int *row_offsets;
     double *values;
-    size_t *column_indices;
+    int *column_indices;
 };
 
 typedef struct csr csr;
